+++ dila/frontend/flask/forms.py
@@ -8,7 +8,7 @@ class PoFileUpload(flask_wtf.FlaskForm):
 
 
 class TranslationForm(flask_wtf.FlaskForm):
-    translation = wtforms.StringField()
+    translation = wtforms.TextAreaField(render_kw={'rows': 5, 'cols': 90})
 
 
 class NewResourceForm(flask_wtf.FlaskForm):