+++ tests/test_views.py
@@ -314,8 +314,8 @@ def test_get_translation_form(get_translated_string, flask_client):
         '1',
     )
     response = flask_client.get('/lang/pl/edit/34/')
-    assert re.search('<input class="[^"]*" id="translation" name="translation" type="text" value="translation-x">',
-                     response.data.decode())
+    assert re.search('<textarea class="[^"]*" cols="[^"]*" id="translation" name="translation" rows="[^"]*">'
+                     'translation-x</textarea>', response.data.decode())
     assert 'action="/lang/pl/edit/34/"' in response.data.decode()